+++ src/controllers/users/deleteUser.js
