+++ src/routes/contacts.js
@@ -1,19 +1,21 @@
 import express from 'express';
 
-import { createContacts } from '../controllers/contacts/CreateContacts';
+import { createContacts } from '../controllers/contacts/createContacts';
 
-import { getSpecificContacts, getAllContacts } from '../controllers/contacts/GetContacts';
+import { getSpecificContacts, getAllContacts } from '../controllers/contacts/getContacts';
 
 import { jsonParser } from '../middleware/bodyParser';
 
 import Contact from '../helpers/validations/contact';
 
+const { validate } = Contact;
+
 const contactRouter = express.Router();
 
 contactRouter.route('/')
   .get(getAllContacts)
-  .post(jsonParser, Contact.validate, createContacts);
-  
+  .post(jsonParser, validate, createContacts);
+
 contactRouter.route('/:id')
   .get(getSpecificContacts);
